update for building in visual studio after merging 'develop'

+++ src/rtcModule/streamPlayer.h
@@ -164,8 +164,8 @@ public:
             {
                 buffer = webrtc::I420Buffer::Rotate(*buffer, frame.rotation());
             }
-            unsigned short width = buffer->width();
-            unsigned short height = buffer->height();
+            unsigned short width = (unsigned short)buffer->width();
+            unsigned short height = (unsigned short)buffer->height();
             void* frameBuf = mRenderer->getImageBuffer(width, height, userData);
             if (!frameBuf) //image is frozen or app is minimized/covered
                 return;
